requirements.txt
numpy==1.21.4
opencv-python==4.5.4.58
pyserial==3.5
